requirements.txt
Django==3.2.8
djangorestframework==3.12.4
psycopg2-binary==2.9.1
python-decouple==3.5
pandas==1.3.3
